refactor(table-tree): Minor cleanup
