--- examples/data-objects/table-tree/src/Table/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { TableView } from "./tableView.js";
-export { TableDataObject } from "./dataObject.js";
