+++ examples/data-objects/table-tree/src/react/index.ts
@@ -3,4 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export { useTree } from "./utilities.js";
+export { TableView } from "./tableView.js";
