+++ examples/data-objects/table-tree/src/react/utilities.ts
