+++ examples/data-objects/table-tree/src/schema.ts
