+++ examples/data-objects/table-tree/tsconfig.json
@@ -16,5 +16,5 @@
 		"noUncheckedIndexedAccess": false,
 		"exactOptionalPropertyTypes": false,
 	},
-	"include": ["src/**/*"],
+	"include": ["src/**/*", "src/view/utilities.ts"],
 }
